% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochecker-package.r
\docType{package}
\name{geochecker}
\alias{geochecker}
\alias{geochecker-package}
\title{geochecker: Check Geocoded Coordinates in a Shiny Gadget}
\description{
Given a data frame of latitude and longitude coordinates which have been
geocded from place names or addresses, the Shiny gadget in this package lets
you conveniently check the accuracy of the locations.
}

