% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocheck.R
\name{geocheck}
\alias{geocheck}
\title{Check geocoded coordinates}
\usage{
geocheck(data, latitude = NULL, longitude = NULL, checked = "checked")
}
\arguments{
\item{data}{A data frame containing latitude and longitude coordinates, and
possibly other metadata to identify the geocoded place.}
}
\description{
Given a data frame of geocoded coordinates, check their locations on the map
one by one.
}

